Use only tslint for linting .ts files with ALE
By default ALE uses eslint for .ts files(together with prettier and tslint).
If eslint is installed by some module which needs it, ALE picks it up
and uses it, even if no .eslintrc file is present in the root directory.
This causes some unwanted behaviour sometimes and even though ALE docs
say that it is better to use eslint for performance reasons, tslint
still does the job without noticable lag.